add check before removing DATA_DIR for test_modifier unit test

+++ tests/test_utils/conftest.py
@@ -7,4 +7,5 @@ from .test_modifier import DATA_DIR
 
 @pytest.hookimpl(tryfirst=True)
 def pytest_sessionfinish(session, exitstatus):
-    shutil.rmtree(DATA_DIR)
+    if DATA_DIR.exists():
+        shutil.rmtree(DATA_DIR)
